Add some debugging, handle edge cases differently
No longer clip the '' element from the top of the pathelts,
so maybe the as-yet-untried "matches" machinery can use it.
Special-case it so the old way still works.

Add the .META files back in.

Add in special THISQUERY file that shows me the query, vals, and
active list for the query for that directory, for debugging.
I expect this to be huge!